These are no-impact change refactors to get the LayoutExplorer code in shape for extension.
