+++ packages/devtools_app/test/layout_explorer/inspector_data_models_test.dart
@@ -3,8 +3,8 @@
 // found in the LICENSE file.
 
 import 'package:devtools_app/src/inspector/inspector_data_models.dart';
-import 'package:devtools_app/src/inspector/layout_explorer/flex/flex.dart';
-import 'package:devtools_app/src/inspector/layout_explorer/flex/utils.dart';
+import 'package:devtools_app/src/inspector/layout_explorer/ui/theme.dart';
+import 'package:devtools_app/src/math_utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
